[Bug 490695] cleand-up interface create command
in order to simplyfy the mirroring of inteface elemnts as needed for
mapped
subapps the interface create command was cleaned up.
Furthermore a default value for the comment parameter has been
introduced. This
should releave as from many null checks as we are needeing them now.


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=490695
Signed-off: alois.zoitl@gmx.at