[519951] Resource editor is using more from FBNetwork editor
With the new model it was now possible to harmonize the code between
FBNetworkEditor (App, Subapp, CFB) and resrouce editor. In this commit
a first set of harmonization has been performed. This greatly reduced
the code required for the resource editor and what it even more 
important: it resulted in a common look and feal especially context 
menus and single source of code. With this new features for the 
FBNeditor should automatically be available for the resource editor 
as well.


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=519951
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>