[512067] Fixed fb tester
The main reason for the tester problem was that the activate function
of the TestEdiPart did not call super.activate and therefore the
editpart was not in active state which prevented an update of the
monitored values. 

Furthermore i removed the printlns to reduce the output in the console
during testing.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=512067
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>