[527152] drop fbs from palette into cfb network
fixed drag and drop from the composite's fb network editor's palette. 
The reason was that a new pallete viewer id was introduced for allowing
that subapps are shown in application and subapplicatoin editors but not
in composite fbs. But this id was not correctly used in the dnd setup.
This has been fixed in this commit.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=527152
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>