[512078] first version which is deploying subapp internals
With this commit the subapp hierachy is correctly parsed and the
contained FBs and subapps are added to the deployment data. Furthermore 
Subappp internal connections are also correctly handled.

As part of this work it has been identified that subapp parsing needed
an adjustment such that if no subappnetwork element is present the 
subappnetwork element is set to null. This may needs further
investigation.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=512078
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>