[527770] reimplemented to open opposite resource feature
With the new virtual IO implementation in this commit also the 
functionality of navigating to the other and of connections crossing
resource boundaries was implemented. 

As part of this work also the open infrastructure for 4diac model
element was cleaned-up and lots of duplicated code was moved in helper
or base classes. This should greatly reduce the effort for implementing
new openers.


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=527770
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>