[508190] Add DMG generation and executable signing in tycho build
In order to adress the issue of running 4diac on MACs and also windows I
enabled three plugins which when running on the Eclipse build
infrastructure will not only sign the pluings and the windows and mac
executable but also generate a DMG package for Macs. This should solve
the issue.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=508190
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>