[531235] added workspace job to load system in resource change listener
In the resource change listener we can not directly load a system as
there markers are deleted. This results in workspace changes which are
not allowed in resoruce change listeners. Not all load systems where
part of a workspace job. In order to clean-up the code and fixe the
issue the workspace job creation and handling has been moved to the load
systems method.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=531235
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>