[532493] replaced getCommandStack.execute with executeCommand
in the propety sheets in many places the cmd stack is set to zero to
indicate that a value was updated. In not all places a null check was
performed. This lead to the cause reported here. In order to halde this
with this commit the already available method executeCommand which is
doing these checks is used.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=532493
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>