[536329] reworked new type wizard to better show types
Experience with the new type wizard has shown that not only file
extension is disturbing but the overall concept with the combo box.
Therefore with this commit the type combo box has been replaced with a
type list that shows not only the type name but also a description taken
from the file. This reduces the number of required clicks by one and no
default selected type can lead to wrong type creations.

Furthermore the open editor checkbox has been moved to the advanced
section.


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=536329
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>