[528070] allowing also the last resource of a device to be moved
With the fix for Bug 537258 it is now no problem that a device has no
resources at all as now a resource can be easily be created on the rest
of the device's figure. Therefore the check on this has been removed.

Furthermore the resource drag and drop between devices has been expanded
to allow also dropping on the header and interface section of a devices
similar to the fix for Bug 537258.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=528070
Change-Id: I714e5130f171a40da9c5dc9669ad9b62cfc7c741
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>