[497979] reworked connection routing to correctly store bendpoints
with this fix now the adjustable connection router correctly generates
the routing information according to IEC 61499-2 and this is also now
stored correctly in the xml file.

Furthermore the whole router infrastructure was reworked to allow
finally fixing also the zoom problem and also give the posiblity to
introduce cascading routers for handling connection overlap issues.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=497979
Change-Id: Ie64230e3adc65374ae0beadcf4a5d4eeb7c8a1c8
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>