[539039] changed action list to a table view
with this change the action list in the state's property sheet is now a
table with two columns: one for the algorithm one for the table. This
allows not only as it was to create and reorder the actions but also to
edit the algorithms and output events of actions.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=539039
Change-Id: I7a454da553f3e80907cfec3954fb36918c4b1e2e
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>