[536266] automatically deactivating monitoring for deployment commands
with this fix it is now possible to have monitoring running and invoke
mgm commands on a device or resources (e.g., delete resource, kill
device, online create) as monitoring is automatically checked.

furthermore the abstractdeployment command has been cleaned up and an
interface introduced that makes it easier to use an
IDeviceManagementInteractor in a try with resources statement.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=536266
Change-Id: I16b3f33ffa6b1154dbc69106aade0fc5667e2c0f
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>