[536266] deactivating monitoring during normal deployment
with this fix now also during normal deployment the monitoring is
automatically deactivated and reactivated after the deployment. This can
be tested by downloading one resource monitor this part and then
download the second one.

It will get even more helpful when the deployment also checks on the
resources in a device.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=536266
Change-Id: If70ff1eea7676d5934afe9ba979dc12e1d975e76
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>