[539254] added first version of fully clean device
with this commit it is now possible to fully clean a device. Thereby
4diac-ide uses the new query resources feature to get a list of all
currently instantiated resources. Before deleting the user is presented
with the resource list to delete and asked for confirmation.

Furthermore some minor code cleanup in the deployment commands is
performed.

Also the debug menu is now put always after the run menu entry in the
main menu.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=539254
Change-Id: Id948f0d6f174f6738dcb0811c9822e0595affb43
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>