[509426] fixed context menu for clear focus on predecessor
With this fix clear focus on predecessor is available in all fbnetwork
editors as context menu as well it is activated in the toolbar and the
navigate menu (where it was moved). So that predecessor highlighting can
be deactivated easily.

furthermore with this fix the Focus on Predecessor action was moved to
the new handler framework and added to the navigation menu so that it
can not only be accessed from the context menu.


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=509426
Change-Id: I2ad4ded340f2cd5350377f4edf55b960f0f18006
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>