[514508] Newly implemented to add fb network element to subapp cmd
In this commit to functionality of selected FB Network elements (i.e.,
FBs and SubApps) to a sub-app via drag and drop has been newly
implemented.

In order to reduce code duplication. This new command is now also used
in the new subapplication command to handle the addition of initally
selected elements.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=514508
Change-Id: Ia0c594d5d63e1b23e4a6ca8bd6bfbbdd6146dd4d
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>