Updated target platform to Eclipse 2010-03
After a quick run test 4diac IDE seams to work with the latest plugins.

However Automatic Error Reporting is broken. This measn that developers
will get a warning message when starting with this target platform. This
can be ignored for now. There is already a bug at eclipse Bug 540306
which is hopefully soon fixed.

Also as of Eclipse 2018-12 32bit builds are removed.

Change-Id: I8c80077b31626ca276e040d03be6d6fe95aef29e
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>