[546922] Reworked interface list property sheet of type editors
With this commit the interface list poperty sheets for types and subapps
are harmonzied with the algorithm lists. As the the layout of the
algorithm lists with the + and delete buttons on the top left have
proven to be better usable this has been also done here.

Furthermore the delete and move buttons are only enabled when something
is selected. So that is clearer when they can be used.

Finally some code clean-up removing duplicated code has been done.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=546922
Change-Id: Ib4b14c9ecee6fd03b42fb1efe90d4133a7442842
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>