[546911] first approach for state and action creation with mouse and key
This is a first approach for allowing to create states and actions in
the graphical ecc editor with the mouse and a modifier key (ctrl or
command). That means by pressing ctrl the editor changes to the creation
tool and depending on the mouse position a state or action is created.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=546911
Change-Id: Ie035fe4714f9b1f8ab9829312eca3817a70e43c4
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>