[547583] Fixed interface type change of mapped subapps
When changing the type of an interface element contained in a untyped
mapped subapp the mapped part was not updated. This lead to strange
behavior like connections not being created in the resource.

Also some code clean-up in the interface section property sheets where
done. More is needed see Bug 547663.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=547583
Change-Id: I03f77950c150eb306b8492b8258ecf58673d1432
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>