[549146] fixed initial state can not get actions from property sheet
With this fix the property state checks if the current state is the
initial state and disables the create action button.

Furthermore an additional helper method for ECStates has been introduced
to give a better access to the ECC from it.

Finally as a result of the last merge commit als an update of the
Mainfest file of the fort_ng  plugin had to be don.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=549146
Change-Id: I3df34ff7f9dd8112d0831a3ac0850f1fa71fb097
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>