[532330] Fixed copy paste of FBs
This fix improves two issues in copy paste handling. First of all it
correctly selects the past FBNetwork. Secondly it fixes the update of
the global copy paste handlers. With the last fix now the current active
editor is used as copy and paste target. Before that it was the last
opened which is unfortunately not good and could result that paste was
performed on the wrong editor.


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=532330
Change-Id: I2b35b7cdda5b1b3e4d6ee22f8aa761fb3fcf6370
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>