[549735] Fixed alg. and event handling in action and trans. properties
This fix correctly handles the new order of events and algorithms in the
drop down lists of the action and transition property sheets. Also an
index out of bound fix when selecting none for events and algorithms was
fixed. Introduced for that a new helper method in the model for getting
an algorithm by its name.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=549735
Change-Id: I7241470ee694234447ce8b6106f281d6766bd2b0
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>