[549743] Moved 4diac preferences and use diag. font for resolution calc.
With this commit the import and export is now using the font metrics of
the diagram font for calculating the resolution conversion between IEC
61499-2 XML and 4diac internal coordination system.

In order to allow this the basic 4diac preferences were moved to the
org.eclipse.fordiac.ide.ui plugin. Based on that many old and not needed
references to org.eclipse.fordiac.ide.util were removed.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=549743
Change-Id: Iddb59e2b487050edfcbc5ce3b5db5c2b97b78490
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>