[551042] Fixed checks for connections going and coming from subapps
With this fix now the code for checking if source and destination need
to be swapped (i.e., connections created reverse) is harmonized between
the connection creation and the checks if a connection is allowed. The
more correct code from connection creation has been used as starting
point.

Furthermore some code clean-up in the LinkConstraints class where
performed.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=551042
Change-Id: I7a8f31d32da2f52f62005bdcf41020b19f377f59
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>