[559296] remove data/adapter con check on fb network parsing
In older 4diac versions (pre 1.8) adapter and data connections where
stored in the same list. Therefore we had a check in model parsing to
convert camouflaged data connections to real adapter cons. As these are
really old 4diac IDE versions I think it is save now to remove that
code.


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=559296
Change-Id: I7a16bd214ce5bd3139fd9b5a8be81470242f32e7
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>