[561249] Removed setting of FB corner dimensions in preferences
As changing the corner dimension of FBs leads to graphical issues and as
there is no real advantage for the user to change that value, I suggest
to remove that setting from the Diagram Preferences.
Additionally, the "Maximum Value Label Size" was moved into a group for
a more homogeneous appearance.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=561249
Change-Id: Ic28b3cb87da68d6f8524b608fc28204ae82cc8e1
Signed-off-by: Lisa Sonnleithner <lisa.sonnleithner@jku.at>