[561849] Fixed update issues in ECC
With the performance cleanup of not using EContentAdapter (it is very
greedy) anymore I removed it at to many places in the ECC editor. This
fix brings it back again solving several interface and drawing update
issues.

Furthermore, as the start state can not be changed anymore, some
unnecessary code for managing that in the ECStateEditPart was also
removed.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=561849
Change-Id: Ib0555944c0620e8e844cf22659b843cadf3258d8
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>