[562076] Replaced ECoreUtil.copy in loading code
ECoreutil.copy is very convenient however also very costly. In the xml
loading code it is used mainly for copying interfaces from the types to
the instances. This takes a significant amount of time for loading
systems (ap. 27%).

In this fix a interfacelist copier is introduced optimized for just
copying the interfaces. This resulted in a 25% reduction of loading
times.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=562076
Change-Id: I519e8556a91e478875afb6800474fdd20caff1d1
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>