[562115] Base infra. for a per project palette and data type handling
With this commit the TypeLibrary manages now the palette as well as a
DataTypeLibrary per project. This is the basis for having more sys files
per project and also to have dedicated user defined data types within a
project.

The old DataTypeLibrary instance is kept for smaller migration steps but
marked as deprecated.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=562115
Change-Id: I5b376119ad0607c26e47bf2307fb4a2e044a998b
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>