[562372] Moved file handling and base parsing to CommonElementImporter
With this fix now the CommonElementImporter is in charge of handling any
streams and files. This reduces code duplication and furthermore as now
every importer has a file allows to centrally manage any error markers.

Therefore now the importer will always have a type which should reduce
any strange NPE behavior in the type handling.

bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=562372
Change-Id: I11d9f3ca6e80757fadfff8b6c7f105e0d0268f39