[562572] Fixed application refresh changes in system explorer
With this fix application additions and deletions are correctly updated
in the tree. The main issue was that the parent of the application is
not a direct element in the system explorer tree.

Furthermore to reduce the number of content providers and simplify
maintenance  the system content provider and the system explorer root
content provider where merged.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=562572
Change-Id: I8ae3173d15726c912536b6159e8860c884823105