[562717] DEMUX type
This change introduces the demultiplexer in the model. For mux and demux
a common base class StructManipulator was introduced. Furthermore, it
allows having a block named STRUCT_DEMUX in the library. For this block,
the output has any struct as type - setting a struct for the multiplexer
in the property sheet creates
the required outputs.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=562717
Change-Id: Ie4578804e3b199123c52527eb4c19adb78e81c49
Signed-off-by: Bianca Wiesmayr <bianca.wiesmayr@jku.at>