[563342] Fixed undesired copy/paste and table behavior
This commit adds new commands which copy the whole entry with all of its
values.

Furthermore it introduces compound commands to the copy and paste
mechanism which allows for undo/redo of multi selections.

When an entry is deleted, the entry below it is selected.

If entries get inserted by copy and paste they become selected.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=563342
Change-Id: Ieb18825069265965b920d295f83038f55c87286b
Signed-off-by: Daniel Lindhuber <daniel.lindhuber@jku.at>