[563367] Mapto visibility fix
I could not find a way to disable/enable the ContributionMenu because
the contribution Item has no setEnable method. So I took the approach
from MonitoredSystem Menu where we are showing "None" if the list is
empty, which has an advantage that user can know by looking at "None"
that they have to create System Model before mapping the application.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=563367
Change-Id: I6549bfbc73dc47ec584cdba715d0e3bf77d913e6
Signed-off-by: Muddasir Shakil <muddasir.shakil@jku.at>