[562640] Moved [none] (i.e., Empty lists) to common messages
In order to harmonize the strings the wrongly spelled [None] from
monitoring was replaced with a commons string in the fordiac.ide.ui
plugin. The same string is also used in the ECC editor.

Furthermore the none menu entry in the monitored systems is now disabled
to clearly indicate that it is not clickable.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=562640
Change-Id: Iabfbf930c5873b3a1da42b7a3648883b85a3c9db
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>