[558338] Improved performance of resource change listener
On file copy or system import the resource change listener was handling
system files rather conserveratively. For large systems this could
result in long waiting time. In order to improve this the name stored in
the system file is retrieved and checked. Only if the name is not
correct the file is parsed and the system name and file name are
harmonized.

As further protection the project and system cache is now only accessed
synchronized.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=558338
Change-Id: I7fe33bd9362c5bf779fe3cb8263c04832078f122
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>