[565266] First version of new InterfaceElementSection
This is a first version of a new InterfaceElementSection.
The TypeAndCommentSection now consists of two parts,
a Type Info Section (read only)
and an Instance Info Section where the current value (if possible) can
be set.
Additionally, a button was added to open a Struct or Adapter in an
editor.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=565266
Change-Id: Ife3d6ca3a6491cc08ab7da59d14eb37dfc7a2112
Signed-off-by: Lisa Sonnleithner <lisa.sonnleithner@jku.at>