[560779] Changing calculation of code metrics
Calculation of code metrics is based on a paper, as mention in the bug
report, but the current implementation does not perform the calculation
as proposed in that paper.
This is fixed now. Halstead metrics and cyclomatic complexity are
calculated for each block and then averaged for CFB and apps.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=560779
Change-Id: I4817a36c8aeac9dc66c993c6a0fd89a17a3b39b7
Signed-off-by: Lisa Sonnleithner <lisa.sonnleithner@jku.at>