[567369] Fixed connection anchor position
By defining an odd anchor position size and calculating the middle of
the label both the normal connection anchor as well as the adapter
anchor are now better aligned with the center of the line.

This works best at 100% zoom. On other zoom levels rounding errors my
reduce the alignment quality. Especially the one with the connection.
Here I tried to improve the quality by shifting some calculations.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=567369
Change-Id: I9a2bfaa887b85789c902292e73bd7de9e62bc8d9
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>