[567611] Code generator produces invalid code for FBs with 0 event ins or outs
This patch provides a fix for this, by adding a clause if no withs are
used with any in or output events. In this case the creation of the
array definitions of scm_anEIWith and/or scm_anEOWith are prevented and
the interface spec is populated with nullptrs in their stead.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=567611

Change-Id: I80cb25c148c9161f297670b50f0fd046e84f17d9
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>