[568459] Added more defense checks to SubAppItemProviderForSystem
With this fix SubAppItemProviderForSystem for safety reasons checks if
the fbnetwork is null in all methods working with the FBNetwork of the
subapp. This is needed when the subapp is changed from an untyped to a
typed one.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=568459
Change-Id: Ic4338b5affd8ae8c629e4f95cf542bdc7c24d2a2
(cherry picked from commit 5ed691c31d945bee5ebb6b1fb158f71636073c71)