[568935] Drag and Drop Element from Subapp to Parent Element
It now works to drag a single element from a subapp into the outer
subapp.

It does not work to drag it to the root element.  (EDIT: This works now
for all selected elements)

And it does not work to
drag multiple elements at once. (Edit: This works now)

Edit: 19.11. 17:09 -> positioning needs to be improved

Edit: 24.11 -> the helper methods isNestedInSubapp and getOuterElement
had been moved to the model
The figure which should be dropped to the root edit panel is now placed
at the mouse location

Edit: 30.11 -> Refactored and fixed NPE within
AddElementsToSubAppCommand checkConnections

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=568935
Change-Id: I7ebb442220491b6a0d0be5dd01d50cd800e0d838
Signed-off-by: mjger <michael.jaeger@jku.at>