[569234] Select different build profile
It is now possible to select a maven profile which imports the correct
p2 file from its corresponding folders.

https://maven.apache.org/guides/introduction/introduction-to-profiles.html

This is possible with the command e.g. mvn install -P nightly

It is also possible to select multiple profiles
if no profile is activated, then per default the profile latest will be
selected

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=569234
Change-Id: Id01eaace0252ea350742dd15b24dd6ef1c36c262
Signed-off-by: mjger <michael.jaeger@jku.at>