[569661] First version of interface bar layout aligned to new canvas
With this change the interface bars of subapps, cfbs, and typed subapp
editors are always shown at the left and right border of the editing
area. They always take the full vertical space. If the drawing area
grows they are automatically adjusted.

This change conflicts a bit with the ELK layouter

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=569661
Change-Id: Ifa212cd3605fa37fcde2a835eeb91b62d52939cb
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>