[570309] Added parent check for connection command canExecute
Connections to the interface of subapps was problematic as the wrong
parent FBNetwork was used. With this change the canExecute method checks
and tries to find the right parent. With that now also connections from
and two subapp interfaces inside and outside are fixed.

The open point which I have currently no solution is cut through
connections. From inputs to outputs. These are currently always added to
the outside of subapps.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=570309
Change-Id: I2040e7cc36a0ec6d1eee1ad49ef807aae432999f
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>