[567670] move to parent: fix duplicate name bug
The move to parent feature did not yet consider that in the parent
network an element of the same name could already exist. With the help
of the NameRepository, we now set the correct name. Undo/Redo restores
the names / creates the name again (so the new name is not permanent,
undo will reset the old name).

https://bugs.eclipse.org/bugs/show_bug.cgi?id=567670
Change-Id: I9cf7eb02a24eb7b601b9aa7335c6df3b560fe135
Signed-off-by: Bianca Wiesmayr <bianca.wiesmayr@jku.at>