[570300] Fixed fb selection for double click in system explorer
With the new breadcrumb open action double clicking on an FB did open
the containing editor in the breadcrumb editor. But the FB was not
selected.

For this fix I had to move the AdvancedScrollingGraphicalViewer and the
BreadcrumbUitl classes into model ui. Espically about the first I'm not
very happy. Hope to find a better solution.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=570300
Change-Id: Ic7175b4fa12886a69c70d960e8c355ed8a585ce6
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>