[490601] Introduced position class in model
During the last model clean-up the introduced positionable element had a
x and y child. As most of the time both values have to be updated this
leads to twice the necessary notifications.

With this fix a new object position is introduced with some helper
methods. This allows that the position of PositionableObjects can be
updated with one setter resulting in only one notification. This should
improve the performance when many positions are set at the same time
(e.g., Auto Layout with ELK).


Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=490601
Change-Id: I5d1ba8bba40f9900a87fd87cbb7686d6f4d2d890
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>