[571098] Changed initial position to top left drawing area corner
With the new drawing canvas on opening applications or subapp the scroll
bar is always set to show as top left 0/0. Especially when the content
is also in the negative area this leads to strange look and feel.

With this change now always the top left corner of the drawing area with
a little offset is shown initially. Such the user always knows in which
direction more content can be found.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=571098
Change-Id: Ie06b52be3e4ef95e02d50f7b0210bb1df16c3765