[571624] Removed FB shape border, reduced FB corner radius
Based on the previous commit the shadow was sharpened and with that the
FB border could be dropped. This leads to nicer looking FB shapes,
especially for mapped FBs.

Furthermore as suggested in the bug report the corner radius was reduced
to half to make the fb shape better visible.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=571624
Change-Id: I8303a4d61135050bd82fd05824ce073a9f7c008a
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>