[571818] Fixed selection actions for all editors
As now most graphical editors are either in the bread-crumb or part of a
multipage editor the selection actions are not updated with the default
selection changed method. To take that into account the special
selection changed method was moved from the different editors that
already did it to the diagram editor base classes.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=571818
Change-Id: Icccb8a2835f0a1a00e48cfc6b562119e5abc01af
Signed-off-by: Alois Zoitl <alois.zoitl@gmx.at>