[571095] Quick fixes for Usability
If you connect two FBs and one pin is ANY type the connection will
be coloured in the typed pin's colour.
For example:
ANY to LREAL will be in the LREAL's colour.
LREAL to ANY will be in the LREAL's colour.
ANY_BIT to ANY will be in the ANY_BIT's colour.
ANY to ANY_BIT will be in the ANY_BIT's colour.

Updated copyright header

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=571095
Change-Id: I83a1d42da6d5679c9b2e9ebde84cb4125a5a3a45
Signed-off-by: Lukas Wais <lukas.wais@jku.at>