[572455] added menu entries GoTo, Open Type in Editor
when an FBnetworkelement is selected in the tree, the menu entry
Navigate/Go To selects and reveals this element. When it is
contained in an untyped subapp, the subapp viewer is opened. Also works
in the context menu of the tree.

Open Type in Editor is shown everywhere and opens the type editor for
typed SubApps and all kinds of FBs. Implemented consistently to the Open
Type in Editor buttons that are spread everywhere.

Based on discussions from previous commits renamed BreadcrumbUtil back
to HandlerHelper (it is useful for all editors, not just breadcrumb)

deactivated open action for fbs (except CFBs)

Conflicts:
	plugins/org.eclipse.fordiac.ide.application/src/org/eclipse/fordiac/ide/application/handlers/MoveToParentHandler.java

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=572455
Change-Id: I36668f016780b543bb8f3c42fd4c1ac6e6ca0ab1
Signed-off-by: Bianca Wiesmayr <bianca.wiesmayr@jku.at>