[570295] Instance viewer fixed visibility of commands with new definition
Instance viewer disabled value editing

--

A new definition was created in the plugin.xml to define the edit
parts in the instance viewer.

This new definition was also added to the go to parent command.
When right clicking on a typed sub app in the instance viewer
the go to parent command is now visible.

--

Flatten subapplication, expand subapplication and follow connection
are now using the new definition.

--

Flatten subapplication is not visibile anymore, follow connection
is only visible when clicking on a connection.
In ConnectionEditPart cleaned up code comments for
getDataConnectioncolor()

[Bug] https://bugs.eclipse.org/bugs/show_bug.cgi?id=570295
Conflicts:
	plugins/org.eclipse.fordiac.ide.application/plugin.xml

Change-Id: Idb269f69ccecdc07bf5478967d17293fbaa5f793