[572379] Updated subapp instance viewer to new lazy loading model
With this change also for subapps contained in typed subapps the viewer
is loaded. Furthermore code in cfb viewers and subapp instance has been
harmonized and updated to new lazy loading of type network into
instances.

This change breaks CFB instance viewers. To fix it a change in the model
is required which will be handled in a separate commit.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=572379
Change-Id: I7be8e2c62c658d9da18076ecb3ccb5623fec1480