[572778] Removes cross-references for events and parameters
Errors in the reference resolution for FB call parameters and event made
this feature almost impossible to use, as same named events or data
ins/outs would randomly be linked to one of the FBs having such a named
event/in/out. As there is no easy way to remedy that, the cross
reference are removed, and the exporter is burdened with resolving the
names to model elements.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=572778

Change-Id: I37aafb8c56afe6d283c48d77f54d1bbb891b6911
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>