[570295] Go to parent implemented new adapters from SubappNetworkEditor
When clicking on a typed subapp inside the instance viewer the
goto parent command is working.
-
Removed unnecessary goto parent handler in plugin.xml
Focusing from the instance viewer now works. If you are inside the
instance viewer and you are using the goto parent function the
focusing on the parent is now working.
-
Implemented new added adapters for FBNetworkelement from
SubappNetworkEditor
Works also when clicking in the background
Code clean up

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=570295
Change-Id: Iee545d0131671b0bce701931accdda21adcb4d0c