[5722907] Bugfix for creating two error pins
The error marker list is now checked if the pin is already contained
within the list. If no pin will be found, then a new one will be
created.

The helper methods from the fbNetworkImporter have been moved to the
marker helper, so that they can be reused if an error marker is created
after parsing the xml

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=572907
Change-Id: I29d4870cf1836a868434337b253427c29c359f7d