[573210] Changed MarqueeRectangleFigure
As with the new drawing infrastructure the xor mode of drawing the
original GEF MarqueeRectangleFigure didn't show on Windows.

This commit provides now an own rectangle figure in the selection color
and with thicker boarders. This should draw nicely on Windows.
Furthermore it harmonizes with the other selection feedback figures.

However in order to get this adjustment it was necessary to copy the
MarqueeSelectionTool from GEF to 4diac IDE and adjust it. If possible
try to upstream the core change which allows to adjust the
MarqueeRectangleFigure.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=573210
Change-Id: Ie887bfbe09453daf6c266acfbfb1b99130fad011