[571259] Introduced command stack listener for restoring editors
This commit introduced a command stack listener which is recording the
editor navigation locations for multi-page editors (e.g., bread-crumb
based editors. This allows to restore editor locations for undo and redo
helping the user to orientate on where undo or redo is happening if in
the meantime the editor sub-page has changed.

With this fix also the NewSubappcommand does not need to handle any
editor pages anymore.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=571259
Change-Id: I668c14803bd00c356151f9f7605a75e732441425