[573704] fixed focus handling of type cell editor on Linux
This fix checks if the mouse is in the overall editor area and only if
outside deactivates the cell editor. Otherwise the dialog button is not
working under Linux.

Furthermore some clean-up of the code has been done. As part of that I
tried to minimize the number of times the datatypes are loaded and
sorted. This should now only happen when the editor is activated.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=573704
Change-Id: Ifec2966e2b581f70a3dc3f806c24a9d10dadf135