[573855] Double click in system navigator opens subapp children
When link with editor was deactivated a double click on a untyped subapp
contained in a typed subapp did not open the right breadcrumb and also
not the right type editor tab. With this fix the open action searches
for the first breadcrumb editor if it can not adapt to it.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=573855
Change-Id: I81e1119b604734fca782c782fde5638b92b9f1e2