[573926] Fixed closing direct edit in ec actions
with the fix for showing correct values when in direct edit a value was
invalid the direct edit of ec actions (algs and output events) an error
was thrown when clicking on the background.

This is fixed with this commit and also the implementation is cleaner as
directyl the DirectEditPolicy is used.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=573926
Change-Id: I8e794cb1b5da5595d77cc46f0cc314654d242f22