[574130] correctly creating adapter event names again
With one of the model clean-ups the getName method of the AdapterEvent
class got lost (it was not in the model but manually added code). This
resulted that ECCs with adapter events where not correctly parsed and
that adapter events showed up without the adapter instance name in the
list.

This fix brings back the getName method to the AdatperEvent class but
this time as part of the model.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=574130
Change-Id: I96a1edb0352e0df6f1c307a998ecf30ed34d0901