[574217] closing untyped subapp when subapp is deleted
Extracted the editor closing code in a dedicated generic adapter and
made the subapp handling adapter a dedicated class to get code reuse
between instance viewer and untyped subapp editor.

This closing will also be called when the type of a subapp is changed,
when a untyped subapp is replaced by a typed one, or when the type of a
CFB is changed.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=574217
Change-Id: Ic0192214636c3ac27f432374a1b7ba583c567d75