[574217] fixed issues with detach type and undo of detach type
When a typed subapp was already shown in an instance viewer the detach
from subapp command created a duplicated broken fb network.

Furthermore on undo of detach type the editor was not closed if opened
in the meantime. Also the subapp icon was not correctly updated to the
typed subapp icon.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=574217
Change-Id: Ic71d7aef092aa2577c4e73691403d9190b590715