[573657] Allows implicit cast from untyped (S/D/L)INT to L/REAL
The validator assumed for untyped literals, that it is a literal of the
input type, which lead to the effect, that for L/REALs a L/REAL literal
had to be typed. Implicitly castable INTs were marked as errors. As an
example 5.0 was valid, whereas 5 was not.

This fix checks for L/REAL inputs now it an untyped literal matches INT
format and also accepts all valid values for U(S/D/L)INTs.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=573657

Change-Id: I140f0190f3b90b9a11188ab6f75cd4e0fc6ca4be
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>