[575288] add menu for recording a service sequence
Added menu entry for the Service editor when service sequence is
selected. A dialogue asks for the input event occurrences that should be
processed. Several events are separated by a semicolon. The interpreter
uses the specified events as inputs and then the results are stored in
the selected service sequence.

Conflicts:
	plugins/org.eclipse.fordiac.ide.fbtypeeditor.servicesequence/META-INF/MANIFEST.MF

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=575288
Change-Id: I524d65b4a39af50a5c343b5ed68929f028723d7c
Signed-off-by: Bianca Wiesmayr <bianca.wiesmayr@jku.at>